+++ .gitignore
@@ -3,3 +3,5 @@ venv/
 __pycache__/
 *.json
 *.pptx
+uploads/
+outputs/
